Refactor BehaviorTree.cs to eliminate garbage
Original implementation was generating a small amount of garbage in RunNode.  I contacted the author (https://github.com/7ark) and he kindly brainstormed a solution on the spot.  The allocation was caused by nesting Actions.  The solution was to separate out the nested Actions and implement caching in the calling class.  Thank you for the solution, Cory/7ark!